Fix wrong positiveLong and nonNegativeLong gsoap generation
Fix external dataarray part count
Fix external dataarray part HDF5 file path
Can no more create a partial dataobject which would replace an existing non partial same uuid dataobject.

+++ src/resqml2/AbstractProperty.h
@@ -293,7 +293,7 @@ namespace RESQML2_NS
 		 * 									the dimension of the value array corresponding to the time series
 		 * 									is the number of entry in the series minus one.
 		 */
-		DLL_IMPORT_OR_EXPORT void setTimeIndices(uint64_t startTimeIndex, uint64_t countTimeIndices, EML2_NS::TimeSeries* ts, bool useInterval = false);
+		DLL_IMPORT_OR_EXPORT void setTimeIndices(unsigned int startTimeIndex, unsigned int countTimeIndices, EML2_NS::TimeSeries* ts, bool useInterval = false);
 
 		/**
 		 * Get the time index of this property in its associated time series
